Start removing old dashboards implementation
