+++ lib/sanbase_web/graphql/schema.ex
@@ -41,7 +41,6 @@ defmodule SanbaseWeb.Graphql.Schema do
   import_types(Graphql.ChartConfigurationTypes)
   import_types(Graphql.ClickhouseTypes)
   import_types(Graphql.CommentTypes)
-  import_types(Graphql.DashboardTypes)
   import_types(Graphql.EntityTypes)
   import_types(Graphql.EcosystemTypes)
   import_types(Graphql.ExchangeTypes)
@@ -91,7 +90,6 @@ defmodule SanbaseWeb.Graphql.Schema do
   import_types(Graphql.Schema.BlockchainQueries)
   import_types(Graphql.Schema.ChartConfigurationQueries)
   import_types(Graphql.Schema.CommentQueries)
-  import_types(Graphql.Schema.DashboardQueries)
   import_types(Graphql.Schema.EmailQueries)
   import_types(Graphql.Schema.EcosystemQueries)
   import_types(Graphql.Schema.EntityQueries)
@@ -158,7 +156,6 @@ defmodule SanbaseWeb.Graphql.Schema do
     import_fields(:comment_queries)
     import_fields(:dashboard_queries)
     import_fields(:ecosystem_queries)
-    import_fields(:old_dashboard_queries)
     import_fields(:entity_queries)
     import_fields(:exchange_queries)
     import_fields(:free_form_json_storage_queries)
@@ -202,7 +199,6 @@ defmodule SanbaseWeb.Graphql.Schema do
     import_fields(:blockchain_address_mutations)
     import_fields(:dashboard_mutations)
     import_fields(:dashboard_queries_interaction_mutations)
-    import_fields(:old_dashboard_mutations)
     import_fields(:comment_mutations)
     import_fields(:email_mutations)
     import_fields(:entity_mutations)
