+++ test/sanbase/billing/query_access_level_test.exs
@@ -41,7 +41,6 @@ defmodule Sanbase.Billing.QueryAccessLevelTest do
           :check_annual_discount_eligibility,
           :comments_feed,
           :comments,
-          :compute_raw_clickhouse_query,
           :currencies_market_segments,
           :current_user,
           :dev_activity,
@@ -70,7 +69,6 @@ defmodule Sanbase.Billing.QueryAccessLevelTest do
           :get_access_restrictions,
           :get_auth_sessions,
           :get_available_blockchains,
-          :get_available_clickhouse_tables,
           :get_available_metrics_for_selector,
           :get_available_metrics,
           :get_available_signals,
@@ -81,12 +79,7 @@ defmodule Sanbase.Billing.QueryAccessLevelTest do
           :get_clickhouse_query_execution_stats,
           :get_current_user_submitted_twitter_handles,
           :get_coupon,
-          :get_dashboard_cache,
-          :get_dashboard_panel_cache,
-          :get_dashboard_schema_history_list,
           :get_ecosystems,
-          :get_dashboard_schema_history,
-          :get_dashboard_schema,
           :get_events_for_users,
           :get_free_form_json,
           :get_full_url,
